+++ src/MWC.App/Resources/L.cs
@@ -448,6 +448,8 @@ public static class L
     public static string DetailLinkEstimate  => Get("Detail_LinkEstimate");
     public static string DetailMlo           => Get("Detail_Mlo");
     public static string DetailSignalTrend   => Get("Detail_SignalTrend");
+    public static string DetailVpn           => Get("Detail_Vpn");
+    public static string DetailEapStats      => Get("Detail_EapStats");
     public static string MenuDiagnosticExport => Get("Menu_DiagnosticExport");
     public static string StatusDiagnosticExported(string filename)
         => Format("Status_DiagnosticExported", filename);